pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gfn"
version = "2.5.0"
description = "Manifold: Geometric Intelligence via Symplectic Geodesic Flows | GFN: Geometric Flow Networks"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Manifold Laboratory (Joaquín Stürtz)", email = "ManifoldLabs@proton.me" }
]
requires-python = ">=3.10"
keywords = [
    "artificial-intelligence",
    "deep-learning", 
    "riemannian-geometry", 
    "symplectic-integration",
    "sequence-modeling",
    "geometric-deep-learning"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "pynvml>=11.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0"
]
cuda = [
    "ninja>=1.11.1" 
]

[project.urls]
Homepage = "https://github.com/Manifold-Laboratory/manifold"
Documentation = "https://github.com/Manifold-Laboratory/manifold/tree/main/docs"
Repository = "https://github.com/Manifold-Laboratory/manifold.git"
Issues = "https://github.com/Manifold-Laboratory/manifold/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gfn*", "configs*"]
namespaces = false

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
